+++ fattmerchant-ios-sdk/Cardpresent/MobileReaderDriver.swift
@@ -8,20 +8,6 @@
 
 import Foundation
 
-enum MobileReaderDriverException: OmniException {
-  case couldNotInitialize(detail: String?)
-
-  static var mess: String = "Mobile reader driver error"
-
-  var detail: String? {
-    switch self {
-    case .couldNotInitialize(let d):
-      return d ?? "Could not initialize mobile reader driver"
-    }
-  }
-
-}
-
 protocol MobileReaderDriver {
 
   static var source: String { get }
