+++ fattmerchant-ios-sdk/Cardpresent/Repository/CustomerRepository.swift
@@ -8,24 +8,6 @@
 
 import Foundation
 
-enum CreateCustomerException: OmniException {
-  case customerNameNotSupplied
-  case unknown
-
-  static var mess = "Could not create customer"
-
-  var detail: String? {
-    switch self {
-    case .customerNameNotSupplied:
-      return "Customer name is required"
-
-    default:
-      return "Unknown error creating customer"
-    }
-  }
-
-}
-
 class CustomerRepository: ModelRepository {
   typealias OmniModel = Customer
   var omniApi: OmniApi
