+++ fattmerchant-ios-sdk/Cardpresent/Repository/InvoiceRepository.swift
@@ -8,20 +8,6 @@
 
 import Foundation
 
-enum CreateInvoiceException: OmniException {
-  case Something(String)
-
-  static var mess = "Could not create customer"
-
-  var detail: String? {
-    switch self {
-    case .Something(let desc):
-      return desc
-    }
-  }
-
-}
-
 class InvoiceRepository: ModelRepository {
   typealias OmniModel = Invoice
   var omniApi: OmniApi
