+++ fattmerchant-ios-sdk/Cardpresent/Repository/PaymentMethodRepository.swift
@@ -8,20 +8,6 @@
 
 import Foundation
 
-enum CreatePaymentMethodException: OmniException {
-  case Something(String)
-
-  static var mess = "Could not create payment method"
-
-  var detail: String? {
-    switch self {
-    case .Something(let desc):
-      return desc
-    }
-  }
-
-}
-
 class PaymentMethodRepository: ModelRepository {
   typealias OmniModel = PaymentMethod
   var omniApi: OmniApi
