+++ fattmerchant-ios-sdk/Cardpresent/Usecase/CancelCurrentTransaction.swift
@@ -8,23 +8,6 @@
 
 import Foundation
 
-enum CancelCurrentTransactionException: OmniException {
-  static var mess: String = "Could not cancel current transaction"
-
-  case noTransactionToCancel
-  case unknown
-
-  var detail: String? {
-    switch self {
-    case .noTransactionToCancel:
-      return "There is no transaction to cancel"
-
-    case .unknown:
-      return "Unkown error"
-    }
-  }
-}
-
 class CancelCurrentTransaction {
   internal var mobileReaderDriverRepository: MobileReaderDriverRepository
