+++ fattmerchant-ios-sdk/Cardpresent/Usecase/CapturePreauthTransaction.swift
@@ -8,23 +8,6 @@
 
 import Foundation
 
-enum CapturePreauthTransactionException: OmniException {
-  static var mess: String = "Could not capture funds"
-
-  case errorCapturing(detail: String?)
-  case captureAmountExceedsAuth(detail: String)
-
-  var detail: String? {
-    switch self {
-    case .errorCapturing(let d):
-      return d ?? "Unknown error"
-
-    case .captureAmountExceedsAuth(let d):
-      return d
-    }
-  }
-}
-
 class CapturePreauthTransaction {
 
   typealias Exception = RefundException
