+++ fattmerchant-ios-sdk/Cardpresent/Usecase/DisconnectMobileReader.swift
@@ -8,12 +8,6 @@
 
 import Foundation
 
-enum DisconnectMobileReaderException: OmniException {
-  static var mess: String = "Could not disconnect mobile reader"
-
-  case driverNotFound
-}
-
 class DisconnectMobileReader {
   var mobileReaderDriverRepository: MobileReaderDriverRepository
   var mobileReader: MobileReader
