+++ fattmerchant-ios-sdk/Cardpresent/Usecase/TokenizePaymentMethod.swift
@@ -8,15 +8,6 @@
 
 import Foundation
 
-enum TokenizePaymentMethodException: OmniException {
-  static var mess: String = "Could not tokenize payment method"
-
-  case tokenizationError
-  case couldNotParsePaymentMethodError
-  case merchantMissingHostedPaymentsToken
-  case noCreditCardOrBankAccountSupplied
-}
-
 /// Tokenizes a payment method
 class TokenizePaymentMethod {
