+++ fattmerchant-ios-sdk/Delegates/MobileReaderConnectionStatusDelegate.swift
