+++ fattmerchant-ios-sdk/Exceptions/ConnectMobileReaderException.swift
@@ -1,0 +1,4 @@
+enum ConnectMobileReaderException: OmniException {
+  static var mess: String = "Could not connect mobile reader"
+  case couldNotConnectMobileReader(reader: MobileReader)
+}
