+++ fattmerchant-ios-sdk/Exceptions/CreateCustomerException.swift
@@ -1,0 +1,17 @@
+enum CreateCustomerException: OmniException {
+  case customerNameNotSupplied
+  case unknown
+
+  static var mess = "Could not create customer"
+
+  var detail: String? {
+    switch self {
+    case .customerNameNotSupplied:
+      return "Customer name is required"
+
+    default:
+      return "Unknown error creating customer"
+    }
+  }
+
+}
