+++ fattmerchant-ios-sdk/Exceptions/CreateInvoiceException.swift
@@ -1,0 +1,13 @@
+enum CreateInvoiceException: OmniException {
+  case Something(String)
+
+  static var mess = "Could not create customer"
+
+  var detail: String? {
+    switch self {
+    case .Something(let desc):
+      return desc
+    }
+  }
+
+}
