+++ fattmerchant-ios-sdk/Exceptions/DisconnectMobileReaderException.swift
@@ -1,0 +1,5 @@
+enum DisconnectMobileReaderException: OmniException {
+  static var mess: String = "Could not disconnect mobile reader"
+
+  case driverNotFound
+}
