+++ fattmerchant-ios-sdk/Exceptions/GetConnectedMobileReaderException.swift
@@ -1,0 +1,5 @@
+enum GetConnectedMobileReaderException: OmniException {
+  static var mess: String = "Could not get connected mobile reader"
+
+  case noReaderAvailable
+}
