+++ fattmerchant-ios-sdk/Exceptions/OmniException.swift
@@ -1,11 +1,3 @@
-//
-//  OmniException.swift
-//  fattmerchant-ios-sdk
-//
-//  Created by Tulio Troncoso on 1/13/20.
-//  Copyright © 2020 Fattmerchant. All rights reserved.
-//
-
 import Foundation
 
 /// An error in the Omni platform
