+++ fattmerchant-ios-sdk/Exceptions/SearchForReadersException.swift
@@ -1,0 +1,4 @@
+enum SearchForReadersException: OmniException {
+  static var mess: String = "Couldnt find mobile readers"
+  case noMobileReaderAvailable
+}
