+++ fattmerchant-ios-sdk/Exceptions/VoidTransactionException.swift
@@ -1,0 +1,5 @@
+enum VoidTransactionException: OmniException {
+  static var mess: String = "Could not capture funds"
+
+  case errorCapturing
+}
