+++ fattmerchant-ios-sdk/Services/CardReader/CardReaderService.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol CardReaderService {
+  
+}
